CefNet.Avalonia.WebView now creates browser when attached to logical tree.
This moves browser creation to an earlier point during initialisation
and reduces likelihood of issue #92 manifesting:
While it is probably still possible to force this, regular usage in Avalonia
should now always result in a properly initialised browser before first
navigation, etc. calls are made to the WebView.

Additionally, WebView.OnCreateBrowser now has an overload that accepts a Window object.